Refactor Argo rollout details to utilize new endpoint functions
- Updated the ArgoRolloutDetails component to use the newly defined abortRollout and retryRollout functions from argo-rollout-endpoints.
- Renamed local functions for aborting and retrying rollouts to handleAbortRollout and handleRetryRollout for clarity.
- Adjusted button click handlers to reflect the new function names, improving code readability and maintainability.
- Enhanced the promotion logic by integrating validation and patching functionalities for rollout promotions.
